Change behavior of Get() to propogate err for undefined vars
Liquid allows calling undefined variables and getting `nil` in the
reference implementation.  Originally Get() followed this rule and
returned `nil` if the key was not previously set.  Instead, this
propogates a ErrVarNotFound error instead.  This allows the caller to
decide what they want to do.

This breaks 1:1 method compatibility with the Ruby impl.  But it leads
to a cleaner API design Go-wise.